+++ egui_glium/src/painter.rs
@@ -241,6 +241,23 @@ impl Painter {
         id
     }
 
+    #[deprecated = 'Use NativeTexture::register_native_texture` instead]
+    pub fn register_glium_texture(
+        &mut self,
+        texture: glium::texture::SrgbTexture2d,
+    ) -> egui::TextureId {
+        let id = self.alloc_user_texture();
+        if let egui::TextureId::User(id) = id {
+            if let Some(Some(user_texture)) = self.user_textures.get_mut(id as usize) {
+                *user_texture = UserTexture {
+                    pixels: vec![],
+                    gl_texture: Some(texture),
+                }
+            }
+        }
+        id
+    }
+
     pub fn set_user_texture(
         &mut self,
         id: egui::TextureId,